Remove ExecStartPre as it serves no purpose.
The usage of ExecStartPre was (as far as I remember) adopted because we
in sysv init scripts do a configtest before restart, and this looked
like the natural way to get the same functionality.

Systemd doesn't do this, if you ask for a restart it doesn't do any
checks before stopping the daemon.

As such this is just extra time spent starting up, as well as being very
noisy because we dump a lot of C code into the journal.

This was originally researched and removed by Federico Schwindt in
redhat/varnish.service inside Varnish-Cache.git, but the intent and
change was lost in the subsequent merging of packaging repositories.