Avoid symlink name influencing pidfile name.
Fix by hardcoding as suggested by Federico Schwindt.

This change also reverts the pid file name change that was an uninteded
side effect of a59b3bdd8 in Nov 2015. That commit renamed the pid file
from varnishd.pid to varnish.pid.

From a maintenance branch perspective it is not a good thing to change
these names, but I think going back to the original is the least painful way
moving forward. We'll have systems with sysv init for many years still.

Fixes: #22