Be firmer on upgrade.
Adding Conflict: with libvarnishapi1 makes dpkg -i require extra options
to uninstall, so avoiding that for now.

The current behaviour will upgrade from 4.1.3 without any issues.

Only downside is that the package is left installed on the system, but
most of the files from it is "taken over" by varnish.deb.

This is all very much debian magic to me, but appears to work.

It may be that packages with Conflicts: libvarnishapi1 works without any
extra options if installing with apt and not dpkg -i.