Remove syntax check on restart/reload
Due to https://github.com/varnishcache/varnish-cache/issues/2316
we cannot reliably config-test the VCL in the init-script.
If users want to configtest the VCL in advance of a restart, this
can be done with loading the new VCL into the running varnish, or
running ``varnishd -C`` with a different option-set that will not
conflict with the running instance(s).