New RPM provides for the varnishd and VMODs
The find-provides script takes the original script and forwards almost
all lines to the original find-provides command provided by RPM. In
order for this to work, %global needs to be used instead of %define to
ensure immediate expansion and avoid infinite recursion.

VMODs are not forwared to find-provides because it matters little to
advertise them as shared objects in the traditional sense.

The output looks like this:

    vmod(blob)
    vmod(directors)
    vmod(purge)
    vmod(std)
    vmod(vtc)
    varnishd(abi) = <sha1>
    varnishd(vrt) = x.y

Partially fixes #8 because in order to make this useful for VMODs
without forcing package maintainers to manually require either the
exact ABI or a given VRT version, a means for automatic requires
for VMOD builds is missing. This doesn't address the NOGIT case,
but we don't support packaging not originating from a git commit.

This would make #92 obsolete once something similar is done on the
DPKG side.