New DPKG provides for the varnishd ABI
This change aligns DPKG provides to their RPM counterparts with one
notable exception: the ABI being a hash, it may start with a letter
and go against DPKG's policy. Since it's only used for strict equality
this is not a problem.

The new provides look like this:

    Provides:
     libvarnishapi1,
     varnish,
     varnishd-abi-sha1
     varnishd-vrt (= x.y)

Partially fixes #8 because in order to make this useful for VMODs
without forcing package maintainers to manually require either the
exact ABI or a given VRT version, a means for automatic requires
for VMOD builds is missing. This doesn't address the NOGIT case,
but we don't support packaging not originating from a git commit.

Unlike RPM packaging, this does not list installed VMODs, as I'm
not certain it can be done in the first place (probably though).

Refs c6d66785a4e9b4827a81d498500845dca1f9f388
Closes #92