Switch RPM provides from bitful to archful
While ultimately the goal is to automate virtual provides and requires
to the extent of out-of-tree modules, considering that for now package
maintainers would have to manually adjust their requirements, they
couldn't easily manage either 32bit or 64bit dependencies in a single
Requires tag. While we don't ship 32bit packages, distributions like
Fedora do.

Now a user can easily target the right architecture on a multilib system
with a single dependency:

    Requires: varnishd(abi)%{?_isa} = sha1

This is already part of the upgrade notes I'm redacting and I will amend
the wiki accordingly.