Fix case where admin listen address is empty, and secret is set
Imagine all other options being valid, including having a secret file
set, but we've purposefully left VARNISH_ADMIN_LISTEN_ADDRESS as an
empty string.  (As we're want to do in a local testing VM, while
building tooling around Varnish.)

When we go issue a reload, it fails: "Authentication Required".

This patch ensures that VARNISH_SECRET_FILE is respected, even when we
allow a defaulting behavior for VARNISH_ADMIN_LISTEN_ADDRESS.