Bring backward compatibility back with vcl.list
In current trunk or soon-to-be Varnish 6.2 the format of vcl.list
changed and broke the varnishreload script. Because the temperature
and state column was split into two columns, field offsets became
off by one.

The forward slash character previously used inside the single column
is now part of the field separator when the output of vcl.list is
handled by awk, working for both Varnish 6.2 and previously supported
versions.

Refs #121