rpm: Rely on the configure step to find doc tools
There was already a hint in b2ea64fd1167337d7354856fea06303a5a6fb794
that we lack some packaging stability from the platforms we support
and this is another occurrence. The configure script of a supported
Varnish release will find the tools we need.

This enables build support for el8 in addition to el7.