rpm: Require redhat-rpm-config at run time for el8
Starting with el8 annobin is used to annotate ELF binaries and this is
set up via the %configure RPM macro. As a result annobin becomes a
requirement of the default cc_command in varnishd but it relies on GCC
spec files that are provided by redhat-rpm-config.

So, similar to how we explicitly require GCC at run time we now also
need to bring the annobin setup.