varnishreload: Add a prefix option
On a system where the active VCL's role is merely to switch to a
different label, each tenant might reload its VCL with a dedicated
prefix. Discarding only has an effect on same-prefix reloads.

It can also simply be used for cosmetic reasons to simply replace the
'reload' prefix with something else.