rpm: Ensure %{systemd_requires} availability
RPM macros for systemd used to be part of the main systemd package but
on Fedora they moved to their own subpackage. The %{systemd_requires}
macro is no longer guaranteed to be present in a default build root.

This change enables Fedora Rawhide builds, for a current Rawhide target
of Fedora 36.

Closes #153